+++ gradle/libs.versions.toml
@@ -27,7 +27,6 @@ bytebuddy-agent = { module = "net.bytebuddy:byte-buddy-agent", version.ref = "by
 bytebuddy-core = { module = "net.bytebuddy:byte-buddy", version.ref = "bytebuddy" }
 
 compose-runtime = { module = "androidx.compose.runtime:runtime", version.ref = "compose" }
-composeUi-foundation = { module = "androidx.compose.foundation:foundation" }
 composeUi-material = { module = "androidx.compose.material:material", version.ref = "compose" }
 composeUi-material3-android = { module = "androidx.compose.material3:material3-android", version = "1.3.1" }
 composeUi-material-icons = { module = "androidx.compose.material:material-icons-core", version.ref = "compose" }
