+++ paparazzi-preview-test-junit/src/main/java/app/cash/paparazzi/preview/Snapshot.kt
@@ -1,16 +1,21 @@
 // Copyright Square, Inc.
 package app.cash.paparazzi.preview
 
+import androidx.compose.runtime.Composable
 import app.cash.paparazzi.Paparazzi
 import app.cash.paparazzi.annotations.PaparazziPreviewData
-import com.google.testing.junit.testparameterinjector.TestParameterValuesProvider
+import com.google.testing.junit.testparameterinjector.TestParameter.TestParameterValuesProvider
 
 /**
  * Take a snapshot of the given [previewData].
  */
-public fun Paparazzi.snapshot(previewData: PaparazziPreviewData, name: String? = null) {
+public fun Paparazzi.snapshot(
+  previewData: PaparazziPreviewData,
+  name: String? = null,
+  wrapper: @Composable (@Composable () -> Unit) -> Unit = { it() }
+) {
   when (previewData) {
-    is PaparazziPreviewData.Default -> snapshotDefault(previewData, name)
+    is PaparazziPreviewData.Default -> snapshotDefault(previewData, name, wrapper)
     is PaparazziPreviewData.Empty -> Unit
     is PaparazziPreviewData.Error -> error(previewData.message)
   }
@@ -26,8 +31,6 @@ public fun Paparazzi.snapshot(previewData: PaparazziPreviewData, name: String? =
  */
 public open class PaparazziValuesProvider(
   private val annotations: List<PaparazziPreviewData>
-) : TestParameterValuesProvider() {
-  override fun provideValues(context: Context?): MutableList<PaparazziPreviewData> {
-    return annotations.toMutableList()
-  }
+) : TestParameterValuesProvider {
+  override fun provideValues(): List<PaparazziPreviewData> = annotations
 }
