+++ paparazzi-preview-test-junit/src/main/java/app/cash/paparazzi/preview/Utils.kt
@@ -1,11 +1,16 @@
 // Copyright Square, Inc.
 package app.cash.paparazzi.preview
 
+import androidx.compose.runtime.Composable
 import app.cash.paparazzi.Paparazzi
 import app.cash.paparazzi.annotations.PaparazziPreviewData
 
-internal fun Paparazzi.snapshotDefault(previewData: PaparazziPreviewData.Default, name: String?) {
+internal fun Paparazzi.snapshotDefault(
+  previewData: PaparazziPreviewData.Default,
+  name: String?,
+  wrapper: @Composable (@Composable () -> Unit) -> Unit = { it() }
+) {
   snapshot(name) {
-    previewData.composable()
+    wrapper { previewData.composable() }
   }
 }
